rating-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
pydantic==2.9.2
openpyxl==3.1.5
python-multipart==0.0.12
psycopg2-binary==2.9.10
python-dotenv==1.0.1
tqdm==4.66.5
